pyproject.toml
[project]
name = "bb-notes"
version = "0.1.0"
description = "Apple Notes to Backboard.io sync app with LLM chat"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "backboard-sdk>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

[tool.hatch.build.targets.wheel]
packages = ["."]